+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/DeleteCredentialRequest.java
@@ -11,7 +11,10 @@ import java.util.Objects;
 /** Delete a credential */
 @Generated
 public class DeleteCredentialRequest {
-  /** Force deletion even if there are dependent services. */
+  /**
+   * Force an update even if there are dependent services (when purpose is **SERVICE**) or dependent
+   * external locations and external tables (when purpose is **STORAGE**).
+   */
   @JsonIgnore
   @QueryParam("force")
   private Boolean force;